Write captureGroups() and calculate()
captureGroups() gets passed an operation of 2 values as a string,
the function then captures the first number, the operator, the second
number and then returns those values on an array.

calculate() gets passed the array previously returned by captureGroups()
and then returns the result of the operation by passing the values of
the array to operate().