Allow only one dot per number in the calc display
Whenever a dot is clicked in the calculator it triggers the click event
listener and then it runs addDotToDisplayInput() on line 96. This function
checks if by adding a dot to display it will match a pattern that only allows
one dot per number. If it does the dot will be added, if not it won't.